requirements.txt
aiohttp
contextily
faker
faker-wifi-essid
geopandas
haversine
matplotlib
python-dotenv